requirements.txt
gitpython
PyYAML
requests